[parser] chore: move production types into its own file
